\name{aligned}
\alias{aligned}
\docType{data}
\title{ sample data after alignment }
\description{
  This is the aligned feature data from 4 sample LC/MS profiles.
}
\usage{
data(aligned)
}
\format{
  The format is:
List of 4
 $ aligned.ftrs: A matrix with m/z, median elution time, and signal strength in each profile.
 $ pk.times    : A matrix with m/z, median elution time, and elution time in each profile.
 $ mz.tol      : the m/z tolerance level in alignment
 $ chr.tol     : the retention time tolerance level in alignment
}
\details{}
\source{Dean Jones lab. Emory University, School of Medicine.}
\references{}
\examples{
data(aligned)
}
\keyword{datasets}
