\name{apLCMS-package}
\alias{apLCMS-package}
\alias{apLCMS}
\docType{package}
\title{
Adaptive processing of LC/MS data
}
\description{
The package generates a feature table from a batch of LC/MS spectra. It finds m/z and retention time tolerance levels from the data. A run-filter is used to detect peaks and remove noise. Non-parametric statistical methods are used to find-tune peak selection and grouping. After retention time correction, a feature table is generated by aligning peaks across spectra. 
}
\author{
Tianwei Yu <tyu8@emory.edu>
Maintainer: Tianwei Yu <tyu8@emory.edu>
}
\references{
Bioinformatics. 25(15):1930-36. 
BMC Bioinformatics. 11:559.
J. Proteome Res. 12(3):1419-27.
}
\keyword{ package }