% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.seq.3.R
\name{combine.seq.3}
\alias{combine.seq.3}
\title{An internal function.}
\usage{
combine.seq.3(features)
}
\arguments{
\item{features}{dataframe of retention time, m/z ratio, signal strength.}
}
\value{
returns
\itemize{
  \item masses - m/z ratio
  \item rt - retention time
  \item intensi - signal strength
}
}
\description{
This is a internal function.
}
\examples{
combine.seq.3(table)
}
