% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_clusters.R
\name{compute_clusters}
\alias{compute_clusters}
\title{Compute clusters of mz and rt and assign cluster id to individual features.}
\usage{
compute_clusters(
  feature_tables,
  mz_tol_relative,
  mz_tol_absolute,
  mz_max_diff,
  rt_tol_relative,
  do.plot,
  sample_names = NA
)
}
\arguments{
\item{feature_tables}{list List of tibbles containing features.
Each tibble should have columns [mz, rt, area].}

\item{mz_tol_relative}{float Relative mz tolerance to use for grouping features.}

\item{mz_tol_absolute}{float Absolute mz tolerance to use for grouping features.}

\item{mz_max_diff}{float Maximum difference between featuure mz values to belong to the same cluster.}

\item{rt_tol_relative}{float Relative retention time tolerance to use for grouping features.}

\item{do.plot}{bool Plot graphics or not.}

\item{sample_names}{list List of sample names.}
}
\value{
Returns a list with following items:
\itemize{
  \item feature_tables - list - Feature tables with added columns [sample_id, cluster].
  \item rt_tol_relative - float - Newly determined relative rt tolerance.
  \item mz_tol_relative - float - Newly determined relative mz tolerance.
}
}
\description{
Uses tolerances to group features with mz and rt within the tolerance into clusters,
creating larger features from raw data points. Custom tolerances for mz and rt are
computed based on the given parameters.
}
