\name{features.learn}
\alias{features.learn}
\docType{data}
\title{ Sample feature tables from 4 profiles. The original feature detection is done by machine learning approach. }
\description{
  A list object containing 4 matrices, each of which is the feature table from a profile.
}
\usage{
data(features)
}
\format{
  List object containing multiple matrices. One matrix from each spectrum.
}
\details{
}
\source{
Data from Dean Jones lab, Emory University School of Medicine.
}
\references{
}
\examples{
data(features.learn)
}
\keyword{datasets}
