\name{features}
\alias{features}
\docType{data}
\title{ Sample feature tables from 4 profiles }
\description{
  A list object containing 4 matrices, each of which is the feature table from a profile.
}
\usage{data(features)}
\format{
  List object containing multiple matrices. One matrix from each spectrum.
}
\source{
Data from Dean Jones lab, Emory University School of Medicine.
}
\examples{
data(features)
}
\keyword{datasets}
