\name{features2.learn}
\alias{features2.learn}
\docType{data}
\title{ Feature tables after elution time correction. The original feature detection is done by machine learning approach. }
\description{
  A list object containing 4 matrices, each of which is from a profile. The elution times in the matrices are corrected.
}
\usage{
data(features2)
}
\format{
A list object. Each component of the list is a matrix.
}
\details{
}
\source{
Raw data from Dean Jones lab, Emory School of Medicine.}
\references{
}
\examples{
data(features2.learn)
}
\keyword{datasets}
