\name{features2}
\alias{features2}
\docType{data}
\title{ Feature tables after elution time correction. }
\description{
  A list object containing 4 matrices, each of which is from a profile. The elution times in the matrices are corrected.
}
\usage{data(features2)}
\format{
A list object. Each component of the list is a matrix.
}
\details{
}
\source{
Raw data from Dean Jones lab, Emory School of Medicine.}
\references{
}
\examples{
data(features2)
}
\keyword{datasets}
