% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.seq.3.R
\name{merge.seq.3}
\alias{merge.seq.3}
\title{An internal function.}
\usage{
\method{merge}{seq.3}(a, mz, inte)
}
\arguments{
\item{a}{vector of retention time.}

\item{mz}{vector of m/z ratio.}

\item{inte}{vector of signal strength.}
}
\description{
This is a internal function. It shouldn't be called by the end user.
}
\examples{



}
\references{
Bioinformatics. 25(15):1930-36.  BMC Bioinformatics. 11:559.
}
\author{
Tianwei Yu <tyu8@emory.edu>
}
\keyword{models}
