% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid.R
\name{merge_features_and_known_table}
\alias{merge_features_and_known_table}
\title{A wrapper function to join knowledge from aligned features and known table.}
\usage{
merge_features_and_known_table(
  features,
  known_table,
  match_tol_ppm,
  mz_tol_relative,
  rt_tol_relative,
  from_features_to_known_table = TRUE,
  new_feature_min_count = NA
)
}
\arguments{
\item{features}{A list object with three tibble tables: metadata, intensity, and rt.}

\item{known_table}{A table of known/previously detected peaks.}

\item{match_tol_ppm}{The ppm tolerance to match identified features to known metabolites/features.}

\item{mz_tol_relative}{The m/z tolerance level for peak alignment. The default is NA, which allows the program to search for the 
tolerance level based on the data. This value is expressed as the percentage of the m/z value. This value, multiplied by the m/z 
value, becomes the cutoff level.}

\item{rt_tol_relative}{The retention time tolerance level for peak alignment. The default is NA, which allows the program to search for 
the tolerance level based on the data.}

\item{from_features_to_known_table}{Determines direction of joining; if TRUE, aligned features are joined to known table, vice verse if it is FALSE.}

\item{new_feature_min_count}{The number of profiles a new feature must be present for it to be added to the database.}
}
\value{
Enriched aligned table or known features.
}
\description{
A wrapper function to join knowledge from aligned features and known table.
}
