\name{merge_seq_3}
\alias{merge_seq_3}
\title{
An internal function.
}
\description{
This is a internal function. It shouldn't be called by the end user.
}
\usage{
merge_seq_3(a, mz, inte)
}
\arguments{
  \item{a}{vector of retention time.}
  \item{mz}{vector of m/z ratio.}
  \item{inte}{vector of signal strength.}
}
\references{
Bioinformatics. 25(15):1930-36. 
BMC Bioinformatics. 11:559.
}
\author{ Tianwei Yu <tyu8@emory.edu> }
\keyword{ models }