% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_filter.R
\name{run_filter}
\alias{run_filter}
\title{Continuity index.}
\usage{
run_filter(newprof, min_pres, min_run)
}
\arguments{
\item{newprof}{The matrix containing m/z, retention time, intensity, and EIC label as columns.}

\item{min_pres}{Run filter parameter. The minimum proportion of presence in the time period for a series of signals grouped
by m/z to be considered a peak.}

\item{min_run}{Run filter parameter. The minimum length of elution time for a series of signals grouped by m/z to be considered a peak.}
}
\value{
A list is returned. new_rec - The matrix containing m/z, retention time, intensity, and EIC label as columns after applying the run filter.
}
\description{
Internal function that removes noise in the retention time dimension. It uses continuity index (or "run filter") to select putative peaks from EIC.
}
