% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive.bin.R
\name{adaptive.bin}
\alias{adaptive.bin}
\title{Adaptive binning}
\usage{
adaptive.bin(
  features,
  min_run,
  min_pres,
  mz_tol,
  baseline_correct,
  intensity_weighted = FALSE
)
}
\arguments{
\item{features}{A tibble with columns of m/z, retention time, intensity.}

\item{min_run}{Run filter parameter. The minimum length of elution time for a series of signals grouped by m/z to be 
considered a peak.}

\item{min_pres}{Run filter parameter. The minimum proportion of presence in the time period for a series of signals grouped 
by m/z to be considered a peak.}

\item{mz_tol}{m/z tolerance level for the grouping of data points. This value is expressed as the fraction of the m/z value. 
This value, multiplied by the m/z value, becomes the cutoff level. The recommended value is the machine's nominal accuracy 
level. Divide the ppm value by 1e6. For FTMS, 1e-5 is recommended.}

\item{baseline_correct}{After grouping the observations, the highest intensity in each group is found. If the highest 
is lower than this value, the entire group will be deleted. The default value is NA, in which case the program uses the 
75th percentile of the height of the noise groups.}

\item{intensity_weighted}{Whether to weight the local density by signal intensities.}
}
\value{
A list is returned.
\itemize{
  \item height.rec - The records of the height of each EIC.
  \item masses - The vector of m/z values after binning.
  \item rt - The vector of retention time after binning.
  \item intensi - The vector of intensity values after binning.
  \item grps - The EIC rt, i.e. which EIC each observed data point belongs to.
  \item times - All the unique retention time values, ordered.
  \item mz_tol - The m/z tolerance level.
  \item min.count.run - The minimum number of elution time points for a series of signals grouped by m/z to be considered a peak.
}
}
\description{
This is an internal function. It creates EICs using adaptive binning procedure
}
