% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.time.R
\name{adjust.time}
\alias{adjust.time}
\title{Adjust retention time across spectra.}
\usage{
adjust.time(
  extracted_features,
  mz_tol_relative,
  rt_tol_relative,
  colors = NA,
  do.plot = TRUE
)
}
\arguments{
\item{extracted_features}{A list object. Each component is a matrix which is the output from compute_clusters}

\item{mz_tol_relative}{The m/z tolerance level for peak alignment. This value is expressed as the
percentage of the m/z value. This value, multiplied by the m/z value, becomes the cutoff level.}

\item{rt_tol_relative}{The retention time tolerance level for peak alignment.}

\item{colors}{The vector of colors to be used for the line plots of time adjustments. The default is NA,
in which case the program uses a set of default color set.}

\item{do.plot}{Indicates whether plot should be drawn.}
}
\value{
A list object with the exact same structure as the input object features, i.e. one matrix per profile
 being processed. The only difference this output object has with the input object is that the retention time
 column in each of the matrices is changed to new adjusted values.
}
\description{
This function adjusts the retention time in each LC/MS profile to achieve better between-profile agreement.
}
