% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.align.R
\name{feature.align}
\alias{feature.align}
\title{Align peaks from spectra into a feature table.}
\usage{
feature.align(
  features,
  min_occurrence = 2,
  mz_tol_relative = NA,
  rt_tol_relative = NA,
  mz_max_diff = 1e-04,
  mz_tol_absolute = 0.01,
  do.plot = TRUE,
  sample_names = NA
)
}
\arguments{
\item{features}{A list object. Each component is a matrix which is the output from proc.to.feature().}

\item{min_occurrence}{A feature has to show up in at least this number of profiles to be included in the final result.}

\item{mz_tol_relative}{The m/z tolerance level for peak alignment. The default is NA, which allows the
program to search for the tolerance level based on the data. This value is expressed as the
percentage of the m/z value. This value, multiplied by the m/z value, becomes the cutoff level.}

\item{rt_tol_relative}{The retention time tolerance level for peak alignment. The default is NA, which
allows the program to search for the tolerance level based on the data.}

\item{mz_max_diff}{Argument passed to find.tol(). Consider only m/z diffs smaller than this value.
This is only used when the mz_tol_relative is NA.}

\item{mz_tol_absolute}{As the m/z tolerance is expressed in relative terms (ppm), it may not be suitable
when the m/z range is wide. This parameter limits the tolerance in absolute terms. It mostly
influences feature matching in higher m/z range.}

\item{do.plot}{Indicates whether plot should be drawn.}

\item{sample_names}{Given list of sample names.}
}
\value{
Returns a list object with the following objects in it:
\itemize{
  \item aligned.ftrs - A matrix, with columns of m/z values, elution times, signal strengths in each spectrum.
  \item pk.times - A matrix, with columns of m/z, median elution time, and elution times in each spectrum.
  \item mz.tol - The m/z tolerance used in the alignment.
  \item rt.tol - The elution time tolerance in the alignment.
}
}
\description{
Identifies which of the peaks from the profiles correspond to the same feature.
}
\examples{
data(extracted)
feature.align(extracted, mz_max_diff = 10 * 1e-05, do.plot = FALSE)
}
