% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.tol.R
\name{find.tol}
\alias{find.tol}
\title{An internal function that is not supposed to be directly accessed by the user. Find m/z tolerance level.}
\usage{
find.tol(
  mz,
  mz_max_diff = 1e-04,
  aver.bin.size = 4000,
  min.bins = 50,
  max.bins = 200,
  do.plot = TRUE
)
}
\arguments{
\item{mz}{The vector of observed m/z values.}

\item{mz_max_diff}{Consider only m/z diffs smaller than this value.}

\item{aver.bin.size}{The average bin size to determine the number of equally spaced points in the kernel density estimation.}

\item{min.bins}{the minimum number of bins to use in the kernel density estimation. It overrides aver.bin.size when too few observations are present.}

\item{max.bins}{the maximum number of bins to use in the kernel density estimation. It overrides aver.bin.size when too many observations are present.}

\item{do.plot}{Indicates whether plot should be drawn.}
}
\value{
The tolerance level is returned.
}
\description{
The function finds the tolerance level in m/z from a given vector of observed m/z values.
}
