% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.tol.R
\name{find.tol}
\alias{find.tol}
\title{An internal function that is not supposed to be directly accessed by the
user. Find m/z tolerance level.}
\usage{
find.tol(a, uppermost = 1e-04, aver.bin.size = 4000, min.bins = 50,
  max.bins = 200)
}
\arguments{
\item{a}{The vector of observed m/z values.}

\item{uppermost}{Consider only m/z diffs smaller than this value.}

\item{aver.bin.size}{The average bin size to determine the number of equally
spaced points in the kernel density estimation.}

\item{min.bins}{the minimum number of bins to use in the kernel density
estimation. It overrides aver.bin.size when too few observations are
present.}

\item{max.bins}{the maximum number of bins to use in the kernel density
estimation. It overrides aver.bin.size when too many observations are
present.}
}
\value{
The tolerance level is returned.
}
\description{
The function finds the tolerance level in m/z from a given vector of
observed m/z values.
}
\details{
The method assumes a mixture model: an unknown distribution of m/z
variations in the same peak, and an exponential distribution of between-peak
diffs. The parameter of the exponential distribution is estimated by the
upper 75% of the sorted data, and the cutoff is selected where the density
of the empirical distribution is >1.5 times the density of the exponential
distribution.
}
\examples{


data(prof)
find.tol(prof[[1]][,1])


}
\author{
Tianwei Yu <tyu8@emory.edu>
}
\keyword{models}
