% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.tol.time.R
\name{find.tol.time}
\alias{find.tol.time}
\title{An internal function that find elution time tolerance level.}
\usage{
find.tol.time(
  features,
  number_of_samples,
  mz_tol_relative = 2e-05,
  rt_tol_relative = NA,
  aver.bin.size = 200,
  min.bins = 50,
  max.bins = 100,
  mz_tol_absolute = 0.01,
  max.num.segments = 10000,
  do.plot = TRUE
)
}
\arguments{
\item{features}{A tibble containing features.}

\item{number_of_samples}{The number of spectra in this analysis.}

\item{mz_tol_relative}{m/z tolerance level for the grouping of signals into peaks. This value is expressed as the percentage of the m/z value.
This value, multiplied by the m/z value, becomes the cutoff level.}

\item{rt_tol_relative}{the elution time tolerance. If NA, the function finds the tolerance level first. If a numerical value is given,
the function directly goes to the second step - grouping peaks based on the tolerance.}

\item{aver.bin.size}{The average bin size to determine the number of equally spaced points in the kernel density estimation.}

\item{min.bins}{the minimum number of bins to use in the kernel density estimation. It overrides aver.bin.size when too few observations are present.}

\item{max.bins}{the maximum number of bins to use in the kernel density estimation. It overrides aver.bin.size when too many observations are present.}

\item{mz_tol_absolute}{As the m/z tolerance in alignment is expressed in relative terms (ppm), it may not be suitable when the m/z range is wide.
This parameter limits the tolerance in absolute terms. It mostly influences feature matching in higher m/z range.}

\item{max.num.segments}{the maximum number of segments.}

\item{do.plot}{Indicates whether plot should be drawn.}
}
\value{
A matrix with six columns. Every row corresponds to a peak in one of the spectrum. The columns are: m/z, elution time, spread, signal strength,
 spectrum label, and peak group label. The rows are ordered by the median m/z of each peak group, and with each peak group the rows are ordered
 by the elution time.
}
\description{
This function finds the time tolerance level. Also, it returns the grouping information given the time tolerance.
}
\examples{
find.tol.time(mz, chr, lab, number_of_samples = number_of_samples, mz_tol_relative = mz_tol_relative, mz_tol_absolute = mz_tol_absolute, do.plot = FALSE)
}
