% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.turn.point.R
\name{find.turn.point}
\alias{find.turn.point}
\title{Find peaks and valleys of a curve.}
\usage{
find.turn.point(y)
}
\arguments{
\item{y}{The y values of a curve in x-y plane.}
}
\value{
A list object:
\itemize{
  \item pks - The peak positions
  \item vlys - The valley positions
}
}
\description{
This is an internal function which finds the peaks and valleys of a smooth curve.
}
