% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.lcms.R
\name{load.lcms}
\alias{load.lcms}
\title{Loading LC/MS data.}
\usage{
load.lcms(filename)
}
\arguments{
\item{filename}{The CDF file name.}
}
\value{
A list is returned.
\itemize{
  \item masses - The vector of m/z values.
  \item labels - The vector of retention times.
  \item intensi - The vector of intensity values.
  \item times - The vector of unique time points.
}
}
\description{
This is an internal function. It loads LC/MS data into memory.
}
