% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prof.to.features.R
\name{prof.to.features}
\alias{prof.to.features}
\title{Generate feature table from noise-removed LC/MS profile.}
\usage{
prof.to.features(
  profile,
  bandwidth = 0.5,
  min.bw = NA,
  max.bw = NA,
  sd.cut = c(0.01, 500),
  sigma.ratio.lim = c(0.01, 100),
  shape.model = "bi-Gaussian",
  estim.method = "moment",
  do.plot = TRUE,
  power = 1,
  component.eliminate = 0.01,
  BIC.factor = 2
)
}
\arguments{
\item{profile}{The matrix output from proc.cdf(). It contains columns of m/z value, retention time, intensity and group number.}

\item{bandwidth}{A value between zero and one. Multiplying this value to the length of the signal along the time axis helps
determine the bandwidth in the kernel smoother used for peak identification.}

\item{min.bw}{The minimum bandwidth to use in the kernel smoother.}

\item{max.bw}{The maximum bandwidth to use in the kernel smoother.}

\item{sd.cut}{A vector of two. Features with standard deviation outside the range defined by the two numbers are eliminated.}

\item{sigma.ratio.lim}{A vector of two. It enforces the belief of the range of the ratio between the left-standard deviation
and the right-standard deviation of the bi-Gaussian function used to fit the data.}

\item{shape.model}{The mathematical model for the shape of a peak. There are two choices - "bi-Gaussian" and "Gaussian".
When the peaks are asymmetric, the bi-Gaussian is better. The default is "bi-Gaussian".}

\item{estim.method}{The estimation method for the bi-Gaussian peak model. Two possible values: moment and EM.}

\item{do.plot}{Whether to generate diagnostic plots.}

\item{power}{The power parameter for data transformation when fitting the bi-Gaussian or Gaussian mixture model in an EIC.}

\item{component.eliminate}{In fitting mixture of bi-Gaussian (or Gaussian) model of an EIC, when a component accounts for a
proportion of intensities less than this value, the component will be ignored.}

\item{BIC.factor}{the factor that is multiplied on the number of parameters to modify the BIC criterion. If larger than 1,
models with more peaks are penalized more.}
}
\value{
A matrix is returned. The columns are: m/z value, retention time, spread (standard deviation of the estimated normal
 curve), and estimated total signal strength (total area of the estimated normal curve).
}
\description{
Each LC/MS profile is first processed by the function proc.cdf() to remove noise and reduce data size. A matrix containing m/z
value, retention time, intensity, and group number is output from proc.cdf(). This matrix is then fed to the function
prof.to.features() to generate a feature list. Every detected feature is summarized into a single row in the output matrix from this function.
}
\examples{
prof.to.features(extracted_features, sd.cut = sd_cut, sigma.ratio.lim = sigma_ratio_lim, do.plot = FALSE)
}
