% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm.ridge.R
\name{rm.ridge}
\alias{rm.ridge}
\title{Removing long ridges at the same m/z.}
\usage{
rm.ridge(x, y2, bw)
}
\arguments{
\item{x}{Retention time vector.}

\item{y2}{Intensity vector.}

\item{bw}{Bandwidth for the kernel smoother. A very wide one is used here.}
}
\value{
A vector of intensity value is returned.
}
\description{
This is an internal function. It substracts a background estimated through kernel smoothing when an EIC continuously 
span more than half the retention time range.
}
