% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm.ridge.R
\name{rm.ridge}
\alias{rm.ridge}
\title{Removing long ridges at the same m/z.}
\usage{
rm.ridge(rt, intensity, bw)
}
\arguments{
\item{rt}{Retention time vector.}

\item{intensity}{Intensity vector.}

\item{bw}{Bandwidth for the kernel smoother. A very wide one is used here.}
}
\value{
A vector of intensity values at each rt intervals is returned.
}
\description{
This is an internal function. It substracts a background when an EIC continuously 
span more than half the retention time range. The background is estimated through kernel smoothing.
}
