% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semi.sup.R
\name{semi.sup}
\alias{semi.sup}
\title{Semi-supervised feature detection}
\usage{
semi.sup(
  files,
  folder,
  known.table = NA,
  n.nodes = 4,
  min.exp = 2,
  min.pres = 0.5,
  min.run = 12,
  mz.tol = 1e-05,
  baseline.correct.noise.percentile = 0.05,
  shape.model = "bi-Gaussian",
  BIC.factor = 2,
  baseline.correct = 0,
  peak.estim.method = "moment",
  min.bw = NA,
  max.bw = NA,
  sd.cut = c(0.01, 500),
  sigma.ratio.lim = c(0.01, 100),
  component.eliminate = 0.01,
  moment.power = 1,
  align.mz.tol = NA,
  align.rt.tol = NA,
  max.align.mz.diff = 0.01,
  pre.process = FALSE,
  recover.mz.range = NA,
  recover.rt.range = NA,
  use.observed.range = TRUE,
  match.tol.ppm = NA,
  new.feature.min.count = 2,
  recover.min.count = 3,
  intensity.weighted = FALSE,
  sample_names = NA
)
}
\arguments{
\item{files}{The CDF file names}

\item{folder}{The folder where all CDF files to be processed are located.}

\item{known.table}{A data frame containing the known metabolite ions and previously found features.}

\item{n.nodes}{The number of CPU cores to be used}

\item{min.exp}{If a feature is to be included in the final feature table, it must be present in at least this number of spectra.}

\item{min.pres}{This is a parameter of thr run filter, to be passed to the function proc.cdf().}

\item{min.run}{This is a parameter of thr run filter, to be passed to the function proc.cdf().}

\item{mz.tol}{The user can provide the m/z tolerance level for peak identification. This value is expressed 
as the percentage of the m/z value. This value, multiplied by the m/z value, becomes the cutoff level.}

\item{baseline.correct.noise.percentile}{The perenctile of signal strength of those EIC that don't pass the run filter, 
to be used as the baseline threshold of signal strength. This parameter is passed to proc.cdf()}

\item{shape.model}{The mathematical model for the shape of a peak. There are two choices - bi-Gaussian and Gaussian. 
When the peaks are asymmetric, the bi-Gaussian is better. The default is bi-Gaussian.}

\item{BIC.factor}{the factor that is multiplied on the number of parameters to modify the BIC criterion. If 
larger than 1, models with more peaks are penalized more.}

\item{baseline.correct}{This is a parameter in peak detection. After grouping the observations, the highest observation in 
each group is found. If the highest is lower than this value, the entire group will be deleted. The default value is NA, 
which allows the program to search for the cutoff level.}

\item{peak.estim.method}{the bi-Gaussian peak parameter estimation method, to be passed to subroutine prof.to.features. 
Two possible values: moment and EM.}

\item{min.bw}{The minimum bandwidth in the smoother in prof.to.features().}

\item{max.bw}{The maximum bandwidth in the smoother in prof.to.features().}

\item{sd.cut}{A parameter for the prof.to.features() function. A vector of two. Features with standard deviation outside 
the range defined by the two numbers are eliminated.}

\item{sigma.ratio.lim}{A parameter for the prof.to.features() function. A vector of two. It enforces the belief of 
the range of the ratio between the left-standard deviation and the right-standard deviation of the bi-Gaussian function 
used to fit the data.}

\item{component.eliminate}{In fitting mixture of bi-Gaussian (or Gaussian) model of an EIC, when a component accounts for 
a proportion of intensities less than this value, the component will be ignored.}

\item{moment.power}{The power parameter for data transformation when fitting the bi-Gaussian or Gaussian mixture model in an EIC.}

\item{align.mz.tol}{The user can provide the m/z tolerance level for peak alignment to override the program's selection. 
This value is expressed as the percentage of the m/z value. This value, multiplied by the m/z value, becomes the cutoff level.}

\item{align.rt.tol}{The user can provide the elution time tolerance level to override the program's selection. This value is 
in the same unit as the elution time, normaly seconds.}

\item{max.align.mz.diff}{As the m/z tolerance in alignment is expressed in relative terms (ppm), it may not be suitable 
when the m/z range is wide. This parameter limits the tolerance in absolute terms. It mostly influences feature matching 
in higher m/z range.}

\item{pre.process}{Logical. If true, the program will not perform time correction and alignment. It will only generate peak 
tables for each spectra and save the files. It allows manually dividing the task to multiple machines.}

\item{recover.mz.range}{A parameter of the recover.weaker() function. The m/z around the feature m/z to search for observations. 
The default value is NA, in which case 1.5 times the m/z tolerance in the aligned object will be used.}

\item{recover.rt.range}{A parameter of the recover.weaker() function. The retention time around the feature retention time to 
search for observations. The default value is NA, in which case 0.5 times the retention time tolerance in the aligned object 
will be used.}

\item{use.observed.range}{A parameter of the recover.weaker() function. If the value is TRUE, the actual range of the observed 
locations of the feature in all the spectra will be used.}

\item{match.tol.ppm}{The ppm tolerance to match identified features to known metabolites/features.}

\item{new.feature.min.count}{The number of profiles a new feature must be present for it to be added to the database.}

\item{recover.min.count}{The minimum time point count for a series of point in the EIC for it to be considered a true feature.}

\item{intensity.weighted}{Whether to use intensity to weight mass density estimation.}
}
\value{
A list is returned.
\itemize{
  \item features - A list object, each component of which being the peak table from a single spectrum.
  \item features2 - each component of which being the peak table from a single spectrum, after elution time correction.
  \item aligned.ftrs - Feature table BEFORE weak signal recovery.
  \item pk.times - Table of feature elution time BEFORE weak signal recovery.
  \item final.ftrs - Feature table after weak signal recovery. This is the end product of the function.
  \item final.times - Table of feature elution time after weak signal recovery.
  \item align.mz.tol - The m/z tolerance level in the alignment across spectra, either input from the user or automatically selected when the user input is NA.
  \item align.rt.tol - The retention time tolerance level in the alignment across spectra, either input from the user or automatically selected when the user input is NA.
  \item mz.tol - The input mz.tol value by the user.
  \item updated.known.table - The known table updated using the newly processed data. It should be used for future datasets generated using the same machine and LC column.
  \item ftrs.known.table.pairing - The paring information between the feature table of the current dataset and the known feature table.
}
}
\description{
The semi-supervised procedure utilizes a database of known metabolites and previously detected features to 
identify features in a new dataset. It is recommended ONLY for experienced users. The user may need to construct 
the known feature database.
}
\examples{
semi.sup(files_batch, work_dir, sd.cut = sd.cut, sigma.ratio.lim = sigma.ratio.lim, moment.power = moment.power, min.exp = ceiling(min.within.batch.prop.detect * length(files_batch)))
}
